test: it for user-linked-object
